requirements.txt
python-dotenv
openai
streamlit
requests 
bs4 
pdfkit